# Conflicts:
#	stdlib/public/core/Mirrors.swift.gyb
#	utils/gyb_syntax_support/Node.py
#	utils/gyb_syntax_support/__init__.py

+++ stdlib/public/Darwin/CoreMedia/CMTime.swift
@@ -62,7 +62,7 @@ extension CMTime {
   }
 
   public var seconds: Double {
-    return CMTimeGetSeconds(self) as Double
+    return CMTimeGetSeconds(self)
   }
 
   public func convertScale(_ newTimescale: Int32, method: CMTimeRoundingMethod)
